Reorganized tenant-specific tests into two classes: one with DML and another with DDL.  A couple of DML tests are failing but it's December 31st and I'm getting out of here
