+++ phoenix-core/src/test/java/com/salesforce/phoenix/end2end/CreateTableTest.java
@@ -27,11 +27,6 @@
  ******************************************************************************/
 package com.salesforce.phoenix.end2end;
 
-import static com.salesforce.phoenix.exception.SQLExceptionCode.BASE_TABLE_NOT_TOP_LEVEL;
-import static com.salesforce.phoenix.exception.SQLExceptionCode.TYPE_ID_USED;
-import static com.salesforce.phoenix.util.PhoenixRuntime.TENANT_ID_ATTRIB;
-import static com.salesforce.phoenix.util.TestUtil.TEST_PROPERTIES;
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
@@ -39,59 +34,17 @@ import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
 
-import org.apache.hadoop.hbase.client.HBaseAdmin;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
 import com.salesforce.phoenix.jdbc.PhoenixStatement;
 import com.salesforce.phoenix.query.KeyRange;
 import com.salesforce.phoenix.schema.TableAlreadyExistsException;
-import com.salesforce.phoenix.schema.TableNotFoundException;
 import com.salesforce.phoenix.util.PhoenixRuntime;
 
 public class CreateTableTest  extends BaseClientMangedTimeTest {
-    
-    private static final String TENANT_ID = "abc";
-    private static final String TENANT_TYPE_ID = "111";
-    private static final String TENANT_TABLE_NAME = "TENANT_SPECIFIC_TABLE";
-    private static final String BASE_TABLE_DDL = "CREATE TABLE PARENT_TABLE ( \n" + 
-            "                user VARCHAR ,\n" +
-            "                tenant_id VARCHAR not null,\n" +
-            "                tenant_type_id VARCHAR(3) not null,\n" +
-            "                id INTEGER not null\n" + 
-            "                constraint pk primary key (tenant_id, tenant_type_id, id)) ";
-    private static final String TENANT_TABLE_DDL = "CREATE TABLE " + TENANT_TABLE_NAME + " ( \n" + 
-            "                tenantCol VARCHAR\n" + 
-            "                ) BASE_TABLE='PARENT_TABLE', TENANT_TYPE_ID='" + TENANT_TYPE_ID + "'";
-
-    private List<String> tenantTableNames = new ArrayList<String>();
-    
-    @Before
-    public void clearTenantTableNames() {
-    	tenantTableNames.clear();
-    }
-    
-    /**
-     * {@link BaseClientMangedTimeTest} automatically drops all tables.  This method is here because
-     * tenant-specific tables need to be dropped first.
-     * @throws SQLException
-     */
-    @After
-    public void dropTenantTables() throws SQLException {
-        Properties props = new Properties();
-        props.setProperty(PhoenixRuntime.CURRENT_SCN_ATTRIB, Long.toString(nextTimestamp()));
-        props.setProperty(TENANT_ID_ATTRIB, TENANT_ID);
-        for (String tenantTableName : tenantTableNames) {
-        	Connection conn = DriverManager.getConnection(getUrl(), props);
-            conn.createStatement().execute("DROP TABLE IF EXISTS " + tenantTableName);
-            conn.close();
-        }
-    }
     
     @Test
     public void testStartKeyStopKey() throws SQLException {
@@ -155,126 +108,5 @@ public class CreateTableTest  extends BaseClientMangedTimeTest {
         props.setProperty(PhoenixRuntime.CURRENT_SCN_ATTRIB, Long.toString(ts + 20));
         conn = DriverManager.getConnection(getUrl(), props);
         conn.createStatement().execute("DROP TABLE m_interface_job");
-    }
-    
-    private void createTable(String ddl, String tenantId, long ts) throws Exception {
-        Properties props = new Properties();
-        if (tenantId != null) {
-        	props.setProperty(PhoenixRuntime.TENANT_ID_ATTRIB, tenantId);
-        }
-        props.setProperty(PhoenixRuntime.CURRENT_SCN_ATTRIB, Long.toString(ts));
-        Connection conn = DriverManager.getConnection(getUrl(), props);
-        conn.createStatement().execute(ddl);
-        conn.close();
-    }
-    
-    @Test
-    public void testCreateTenantSpecificTable() throws Exception {
-    	tenantTableNames.add(TENANT_TABLE_NAME);
-        long ts = nextTimestamp();
-        createTable(BASE_TABLE_DDL, null, ts);
-        createTable(TENANT_TABLE_DDL, TENANT_ID, ts+10);
-        // ensure we didn't create a physical HBase table for the tenant-specific table
-        HBaseAdmin admin = driver.getConnectionQueryServices(getUrl(), TEST_PROPERTIES).getAdmin();
-        assertEquals(0, admin.listTables("TENANT_SPECIFIC_TABLE").length);
-    }
-    
-    @Test
-    public void testCreateTenantTableTwice() throws Exception {
-    	tenantTableNames.add(TENANT_TABLE_NAME);
-        long ts = nextTimestamp();
-        createTable(BASE_TABLE_DDL, null, ts);
-        createTable(TENANT_TABLE_DDL, TENANT_ID, ts+10);
-        try {
-        	createTable(TENANT_TABLE_DDL, TENANT_ID, ts+20);
-        	fail();
-        }
-        catch (TableAlreadyExistsException expected) {}
-    }
-    
-    @Test
-    public void testCreateTenantTableBaseTableTopLevel() throws Exception {
-    	tenantTableNames.add(TENANT_TABLE_NAME);
-        long ts = nextTimestamp();
-        createTable(BASE_TABLE_DDL, null, ts);
-        createTable(TENANT_TABLE_DDL, TENANT_ID, ts+10);
-        try {
-        	createTable("CREATE TABLE TENANT_TABLE2 (COL VARCHAR) BASE_TABLE='TENANT_SPECIFIC_TABLE',TENANT_TYPE_ID='aaa'", TENANT_ID, ts+20);
-        	fail();
-        }
-        catch (SQLException expected) {
-            assertEquals(BASE_TABLE_NOT_TOP_LEVEL.getErrorCode(), expected.getErrorCode());
-        }
-    }
-    
-    @Test
-    public void testCreateTenantTableWithDifferentTypeId() throws Exception {
-    	tenantTableNames.add(TENANT_TABLE_NAME);
-        long ts = nextTimestamp();
-        createTable(BASE_TABLE_DDL, null, ts);
-        createTable(TENANT_TABLE_DDL, TENANT_ID, ts+10);
-        try {
-        	createTable(TENANT_TABLE_DDL.replace(TENANT_TYPE_ID, "000"), TENANT_ID, ts+20);
-            fail();
-        }
-        catch (TableAlreadyExistsException expected) {}
-    }
-    
-    @Test
-    public void testCreateTenantTableWithSameTypeId() throws Exception {
-    	tenantTableNames.add(TENANT_TABLE_NAME);
-    	tenantTableNames.add("TENANT_SPECIFIC_TABLE_II");
-        createTable(BASE_TABLE_DDL, null, nextTimestamp());
-        createTable(BASE_TABLE_DDL.replace("PARENT_TABLE", "PARENT_TABLE_II"), null, nextTimestamp());
-        createTable(TENANT_TABLE_DDL, TENANT_ID, nextTimestamp());
-        
-        // Create a tenant table with tenant type id used with a different base table.  This is allowed.
-        createTable("CREATE TABLE TENANT_SPECIFIC_TABLE_II ( \n" + 
-    		    " col VARCHAR\n" + 
-    		    " ) BASE_TABLE='PARENT_TABLE_II', TENANT_TYPE_ID='" + TENANT_TYPE_ID + "'", TENANT_ID, nextTimestamp());
-        
-        try {
-        	// Create a tenant table with tenant type id already used for this base table.  This is not allowed.
-        	createTable("CREATE TABLE TENANT_SPECIFIC_TABLE2 ( \n" + 
-        		    " col VARCHAR\n" + 
-        		    " ) BASE_TABLE='PARENT_TABLE', TENANT_TYPE_ID='" + TENANT_TYPE_ID + "'", TENANT_ID, nextTimestamp());
-        	fail();
-        }
-        catch (SQLException expected) {
-            assertEquals(TYPE_ID_USED.getErrorCode(), expected.getErrorCode());
-        }
-    }
-    
-    @Test(expected=TableNotFoundException.class)
-    public void testDeletionOfParentTableFailsOnTenantSpecificConnection() throws Exception {
-        long ts = nextTimestamp();
-        Properties props = new Properties();
-        props.setProperty(PhoenixRuntime.CURRENT_SCN_ATTRIB, Long.toString(ts));
-        Connection conn = DriverManager.getConnection(getUrl(), props);
-        conn.createStatement().execute("CREATE TABLE PARENT_TABLE ( \n" + 
-                "                user VARCHAR ,\n" + 
-                "                id INTEGER not null primary key desc\n" + 
-                "                ) ");
-        conn.close();
-        
-        props.setProperty(PhoenixRuntime.TENANT_ID_ATTRIB, TENANT_ID); // connection is tenant-specific
-        conn = DriverManager.getConnection(getUrl(), props);
-        conn.createStatement().execute("DROP TABLE PARENT_TABLE");
-        conn.close();
-    }
-    
-    @Test(expected=SQLException.class)
-    public void testCreationOfParentTableFailsOnTenantSpecificConnection() throws Exception {
-        createTable("CREATE TABLE PARENT_TABLE ( \n" + 
-                "                user VARCHAR ,\n" + 
-                "                id INTEGER not null primary key desc\n" + 
-                "                ) ", TENANT_ID, nextTimestamp());
-    }
-    
-    @Test(expected=SQLException.class)
-    public void testCreationOfTenantSpecificTableFailsOnNonTenantSpecificConnection() throws Exception {
-        createTable("CREATE TABLE TENANT_SPECIFIC_TABLE ( \n" + 
-                "                tenantCol VARCHAR \n" + 
-                "                ) BASE_TABLE='PARENT_TABLE', TENANT_TYPE_ID='abc'", null, nextTimestamp());
     }
 }
